- Moving settings into models and rearranging imports
