+++ nh_eobs_mental_health/__init__.py
@@ -1,3 +1,3 @@
 from .models import spell_obs_stop
+from .models import nh_clinical_settings
 from .models import nh_clinical_wardboard
-from . import settings
